requirements.txt
alembic==1.8.1
anyio==3.6.2
bcrypt==4.0.1
certifi==2022.9.24
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.4
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
fastapi==0.87.0
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.1
httptools==0.5.0
httpx==0.23.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
orjson==3.8.2
passlib==1.7.4
psycopg2==2.9.5
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.21.0
typing_extensions==4.4.0
ujson==5.5.0
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
